Add a renderTemplate hook to use existing template to display bookmarks
This does almost what we want, but there is still no content on the
bookmarks index page. The template generates 20 copies of our
component, with the article tags. But checking the data view
in the inspector shows us that each data record has only one
attribute of id -- we don't see the attributes of the url or
description.